refactor. 서비스 로직 리팩토링

+++ src/main/java/com/bomstart/tobyspring/user/service/UserServiceImpl.java
@@ -27,9 +27,11 @@ public class UserServiceImpl implements UserService{
     @Override
     public void createUser(User user) {
         if(userDao.selectUser(user.getId())!=null){
-            System.out.println("이미 존재하는 id 입니다.");
+            System.out.println("이미 등록된 회원입니다.");
+        } else {
+            System.out.println("회원 등록을 완료하였습니다.");
+            this.userDao.createUser(user);
         }
-        this.userDao.createUser(user);
     }
 
     @Override
@@ -40,6 +42,11 @@ public class UserServiceImpl implements UserService{
 
     @Override
     public void deleteUser(String id) {
-        this.userDao.deleteUser(id);
+        if(userDao.selectUser(id)==null){
+            System.out.println("등록되지 않은 회원입니다.");
+        } else {
+            System.out.println("회원을 삭제했습니다.");
+            this.userDao.deleteUser(id);
+        }
     }
 }
